Translate comments in translator.go from Japanese to English
The existing comments in the translator/translator.go file were written in Japanese. In this commit, they have been translated into English to make the code easier to understand for non-Japanese speakers. Now, the structure of the DeepL API response, translation function, and all the step-by-step processes involved are described in English.