Add GitHub workflow for build and release
This commit includes a new GitHub Actions workflow file 'release.yml'. The workflow is triggered on a push event with tags, or via a workflow dispatch. It includes building the project for 'deepl' and 'deepldoc' directories, uploading build artifacts, and applying conditional checks for creating and attaching assets to the GitHub release.