Update README.md to enhance deepldoc documentation
Made changes to the README.md file to clarify the functionality of the deepldoc tool. Enhanced the description to specify that deepldoc also preserves links in your documents and that text wrapped in triple backticks or tilde characters will not be translated. The FAQ section was removed due to the contents of it being redundant or outdated.